chnage the type of rate in Review

+++ userVoice/Model/Review.cs
@@ -11,7 +11,7 @@ namespace userVoice.Model
     public class Review
     {
         [Required]
-        public double Rate { get; set;  }
+        public int Rate { get; set;  }
         [Required]
         public String Content { get; set;  }
         [Required]
